+++ storage_test.go
@@ -1389,7 +1389,7 @@ func (nonStorable) ChildStorables() []atree.Storable {
 	return nil
 }
 
-func (v nonStorable) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
+func (v nonStorable) Storable(_ atree.SlabStorage, _ atree.Address, _ uint32) (atree.Storable, error) {
 	return v, nil
 }
